Centralize discord series stats caching and harden overview cache writes
