+++ api/commands/stats/tests/stats.test.ts
@@ -39,6 +39,7 @@ import { aFakeEnvWith } from "../../../base/fakes/env.fake";
 import { aFakeGuildConfigRow } from "../../../services/database/fakes/database.fake";
 import { EndUserError } from "../../../base/end-user-error";
 import type { MatchPlayer } from "../../../services/halo/types";
+import { getDiscordSeriesStatsCacheKey } from "../../../services/discord/discord-series-stats";
 
 const applicationCommandInteractionStatsNeatQueue: APIApplicationCommandInteraction = {
   ...fakeBaseAPIApplicationCommandInteraction,
@@ -117,15 +118,15 @@ describe("StatsCommand", () => {
   let statsCommand: StatsCommand;
   let services: Services;
   let env: Env;
-  let fetchSpy: MockInstance<typeof globalThis.fetch>;
+  let appDataPutSpy: MockInstance<typeof env.APP_DATA.put>;
   let updateDeferredReplySpy: MockInstance<typeof services.discordService.updateDeferredReply>;
   let updateDeferredReplyWithErrorSpy: MockInstance<typeof services.discordService.updateDeferredReplyWithError>;
 
   beforeEach(() => {
-    services = installFakeServicesWith();
     env = aFakeEnvWith();
+    services = installFakeServicesWith({ env });
     statsCommand = new StatsCommand(services, env);
-    fetchSpy = vi.spyOn(globalThis, "fetch").mockResolvedValue(new Response(null, { status: 200 }));
+    appDataPutSpy = vi.spyOn(env.APP_DATA, "put").mockResolvedValue();
 
     updateDeferredReplySpy = vi.spyOn(services.discordService, "updateDeferredReply").mockResolvedValue(apiMessage);
     updateDeferredReplyWithErrorSpy = vi
@@ -374,12 +375,14 @@ describe("StatsCommand", () => {
         expect(updateDiscordAssociationsSpy).toHaveBeenCalledWith();
       });
 
-      it("warms the discord series stats route after posting embeds", async () => {
+      it("caches resolved discord series stats after posting embeds", async () => {
         await jobToComplete?.();
 
-        expect(fetchSpy).toHaveBeenCalledWith("http://localhost:8787/api/stats/discord/fake-guild-id/777", {
-          method: "GET",
-        });
+        expect(appDataPutSpy).toHaveBeenCalledWith(
+          getDiscordSeriesStatsCacheKey("fake-guild-id", 777),
+          expect.any(String),
+          expect.objectContaining({ expirationTtl: 60 * 60 * 24 }),
+        );
       });
 
       it("calls discordService.updateDeferredReplyWithError with an error when an error is thrown", async () => {
@@ -677,14 +680,16 @@ describe("StatsCommand", () => {
           expect(updateDiscordAssociationsSpy).toHaveBeenCalled();
         });
 
-        it("warms the discord series stats route after posting embeds", async () => {
+        it("caches resolved discord series stats after posting embeds", async () => {
           getMessagesSpy.mockResolvedValue([threadFirstMessage]);
 
           await jobToComplete?.();
 
-          expect(fetchSpy).toHaveBeenCalledWith("http://localhost:8787/api/stats/discord/fake-guild-id/777", {
-            method: "GET",
-          });
+          expect(appDataPutSpy).toHaveBeenCalledWith(
+            getDiscordSeriesStatsCacheKey("fake-guild-id", 777),
+            expect.any(String),
+            expect.objectContaining({ expirationTtl: 60 * 60 * 24 }),
+          );
         });
 
         it("appends previous error data when new error occurs", async () => {
