+++ api/services/neatqueue/tests/neatqueue.test.ts
@@ -645,6 +645,7 @@ describe("NeatQueueService", () => {
       let haloServiceUpdateDiscordAssociationsSpy: MockInstance<typeof haloService.updateDiscordAssociations>;
       let discordServiceStartThreadFromMessageSpy: MockInstance<typeof discordService.startThreadFromMessage>;
       let discordServiceCreateMessageSpy: MockInstance<typeof discordService.createMessage>;
+      let cacheResolvedDiscordSeriesStatsSpy: MockInstance<typeof discordService.cacheResolvedDiscordSeriesStats>;
 
       beforeEach(() => {
         appDataGetSpy = vi.spyOn(env.APP_DATA, "get");
@@ -659,6 +660,9 @@ describe("NeatQueueService", () => {
           }),
         );
         appDataDeleteSpy = vi.spyOn(env.APP_DATA, "delete").mockResolvedValue();
+        cacheResolvedDiscordSeriesStatsSpy = vi
+          .spyOn(discordService, "cacheResolvedDiscordSeriesStats")
+          .mockResolvedValue();
 
         // Mock live tracker to be inactive so tests fall back to timeline-based flow
         vi.spyOn(liveTrackerService, "getTrackerStatus").mockResolvedValue({
@@ -908,6 +912,13 @@ describe("NeatQueueService", () => {
           expect(discordServiceCreateMessageSpy).toHaveBeenCalledTimes(5);
           expect(discordServiceCreateMessageSpy.mock.calls).toMatchSnapshot();
           expect(appDataDeleteSpy).toHaveBeenCalledWith("neatqueue:state:guild-1:2");
+          expect(cacheResolvedDiscordSeriesStatsSpy).toHaveBeenCalledWith(
+            expect.objectContaining({
+              queueNumber: 2,
+              guildId: expect.any(String),
+              matchIds: expect.any(Array),
+            }),
+          );
         });
 
         it("creates the thread/message and posts overviews and game stats, clears timeline", async () => {
@@ -1141,7 +1152,7 @@ describe("NeatQueueService", () => {
 
             await jobToComplete?.();
 
-            expect(logServiceWarnSpy).toHaveBeenCalledExactlyOnceWith(
+            expect(logServiceWarnSpy).toHaveBeenCalledWith(
               error,
               new Map([["reason", "Failed to post series data to thread"]]),
             );
@@ -1266,6 +1277,7 @@ describe("NeatQueueService", () => {
   describe("handleRetry", () => {
     let fakeMessage: APIMessage;
     let fakeErrorEmbed: EndUserError;
+    let cacheResolvedDiscordSeriesStatsSpy: MockInstance<typeof discordService.cacheResolvedDiscordSeriesStats>;
 
     beforeEach(() => {
       fakeMessage = {
@@ -1357,11 +1369,16 @@ describe("NeatQueueService", () => {
 
       vi.spyOn(haloService, "getSeriesFromDiscordQueue").mockResolvedValue([]);
       vi.spyOn(discordService, "editMessage").mockResolvedValue(apiMessage);
+      cacheResolvedDiscordSeriesStatsSpy = vi
+        .spyOn(discordService, "cacheResolvedDiscordSeriesStats")
+        .mockResolvedValue();
     });
 
     it("processes retry for failed series fetch", async () => {
       const getSeriesFromDiscordQueueSpy = vi.spyOn(haloService, "getSeriesFromDiscordQueue");
       const editMessageSpy = vi.spyOn(discordService, "editMessage");
+      const match = Preconditions.checkExists(getMatchStats("d81554d7-ddfe-44da-a6cb-000000000ctf"));
+      getSeriesFromDiscordQueueSpy.mockResolvedValue([match]);
 
       await neatQueueService.handleRetry({
         errorEmbed: fakeErrorEmbed,
@@ -1371,6 +1388,13 @@ describe("NeatQueueService", () => {
 
       expect(getSeriesFromDiscordQueueSpy).toHaveBeenCalled();
       expect(editMessageSpy).toHaveBeenCalled();
+      expect(cacheResolvedDiscordSeriesStatsSpy).toHaveBeenCalledWith(
+        expect.objectContaining({
+          guildId: "guild-123",
+          queueNumber: 1,
+          matchIds: expect.any(Array),
+        }),
+      );
     });
 
     it("handles missing queue message gracefully", async () => {
